Make updating autosave a redux action
And use debounce-action to debounce it.
